Move files back into gitlab system test helpers
