+++ system_tests/gitlab/_helpers/const.py
@@ -7,6 +7,9 @@ from repobee_testhelpers._internal.templates import TEMPLATE_REPOS_DIR
 
 DIR = pathlib.Path(__file__).resolve().parent
 
+VOLUME_DST = "/workdir"
+COVERAGE_VOLUME_DST = "/coverage"
+
 TOKEN = (DIR.parent / "token").read_text(encoding="utf-8").strip()
 OAUTH_USER = "oauth2"
 BASE_DOMAIN = "gitlab.integrationtest.local"
