+++ system_tests/gitlab/_helpers/helpers.py
@@ -1,7 +1,9 @@
 """Helper functions for the integration tests."""
 import os
 import pathlib
+import shlex
 import subprocess
+import sys
 import tempfile
 
 from typing import Optional
@@ -45,6 +47,32 @@ def get_group(group_slug: str, gl: Optional[gitlab.Gitlab] = None):
         for group in gl.groups.list(search=group_slug)
         if group.path == group_slug
     ][0]
+
+
+def run_in_docker_with_coverage(command, extra_args=None):
+    assert extra_args, "extra volume args are required to run with coverage"
+    coverage_command = (
+        "coverage run --branch --append --source _repobee -m " + command
+    )
+    return run_in_docker(coverage_command, extra_args=extra_args)
+
+
+def run_in_docker(command, extra_args=None):
+    extra_args = " ".join(extra_args) if extra_args else ""
+    docker_command = (
+        "docker run {} --net development --rm --name repobee "
+        "repobee:test /bin/sh -c '{}'"
+    ).format(extra_args, command)
+    print(docker_command)
+    proc = subprocess.run(
+        shlex.split(docker_command),
+        stdout=subprocess.PIPE,
+        stderr=subprocess.STDOUT,
+    )
+    print(
+        proc.stdout.decode(sys.getdefaultencoding())
+    )  # for test output on failure
+    return proc
 
 
 def update_repo(repo_name, filename, text):
