+++ system_tests/gitlab/test_gitlab_system.py
@@ -8,12 +8,6 @@ import repobee_plug as plug
 import repobee_testhelpers
 
 from repobee_testhelpers.funcs import hash_directory
-from repobee_testhelpers._internal.docker import (
-    run_in_docker_with_coverage,
-    run_in_docker,
-    VOLUME_DST,
-)
-from repobee_testhelpers._internal.asserts import assert_cloned_repos
 
 import _repobee.ext
 import _repobee.command.peer
@@ -28,6 +22,7 @@ from _helpers.asserts import (
     assert_on_groups,
     assert_issues_exist,
     assert_num_issues,
+    assert_cloned_repos,
 )
 from _helpers.const import (
     BASE_DOMAIN,
@@ -44,8 +39,11 @@ from _helpers.const import (
     MASTER_REPOS_ARG,
     TEMPLATE_ORG_ARG,
     TEACHER,
+    VOLUME_DST,
 )
 from _helpers.helpers import (
+    run_in_docker_with_coverage,
+    run_in_docker,
     api_instance,
     update_repo,
     expected_num_members_group_assertion,
